requirements.txt
requests==2.32.5
pandas==2.3.3
numpy==2.4.0
matplotlib==3.10.8
python-dotenv==1.2.1